+++ CHANGELOG.md
@@ -3,8 +3,10 @@
 
 ## 2020-02-27
 * [Fix] Fix collapset sidebar style
+* [Fix] Fix editable column width `50px -> 70px`
 * [Fix] Changed default font size to 14px -> 16px
 * [Fix] Fix style for Tabs
+* [Add] Add `setReadonly(bool)` for editable fields
 
 
 ## 2020-02-26